src/backend/requirements.txt
Flask==3.1.1
flask-cors==6.0.1
Flask-Limiter==3.12
google-cloud-storage==2.19.0
google-cloud-aiplatform==1.96.0
google-cloud-logging
google-auth==2.40.3
google-api-core==2.25.0
googleapis-common-protos==1.70.0
grpcio==1.73.0
gunicorn
grpcio-status==1.73.0
proto-plus==1.26.1
protobuf
plotly==6.1.2
kaleido==0.2.1
pillow==11.2.1
matplotlib==3.10.3
numpy==2.3.0
Markdown==3.8.2
weasyprint==65.1
cssselect2==0.8.0
tinycss2==1.4.0
pyphen==0.17.2
Jinja2==3.1.6
requests==2.32.3
python-dotenv==1.1.0
setuptools==80.9.0
